fourth commit

+++ src/main/java/com/mycompany/cw/resources/BillingResource.java
@@ -7,6 +7,8 @@ package com.mycompany.cw.resources;
 import javax.ws.rs.Path;
 import com.mycompany.cw.model.Billing;
 import com.mycompany.cw.DAO.BillingDAO;
+import com.mycompany.exception.ResourceNotCreatedException;
+import com.mycompany.exception.ResourceNotFoundException;
 import java.util.ArrayList;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -16,48 +18,75 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+
 /**
  *
  * @author Hashirhalaldeen
  */
 @Path("/Bill")
 public class BillingResource {
+
     private BillingDAO billDAO = new BillingDAO();
-    
+
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public ArrayList<Billing> getAllBills(){
-        return billDAO.getAllBillingRecords();
-    
+    public ArrayList<Billing> getAllBills() {
+         ArrayList<Billing> bills = billDAO.getAllBillingRecords();
+          if (bills != null) {
+            return bills;
+        } else {
+           throw new ResourceNotFoundException("No Bills Found");
+        }
+       
+
     }
-    
+
     @GET
     @Path("/{billId}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Billing getBillingByInvoiceId(@PathParam("billId") String Id){
-        return billDAO.getBillByInvoiceId(Id);
+    public Billing getBillingByInvoiceId(@PathParam("billId") String Id) {
+        Billing bill = billDAO.getBillByInvoiceId(Id);
+         if (bill != null) {
+            return bill;
+        } else {
+           throw new ResourceNotFoundException("Bill with ID: "+Id+" not found");
+        }
+        
     }
-    
+
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
-     public void addBill(Billing Bill){
-         billDAO.addBillingRecord(Bill);
-     }
-    
+    public Response addBill(Billing Bill) {
+        if (Bill.getInvoiceId() == null) {
+
+            throw new ResourceNotCreatedException("Bill cannot be added");
+        }
+        billDAO.addBillingRecord(Bill);
+        return Response.ok().build();
+    }
+
     @PUT
     @Path("/{billId}")
     @Consumes(MediaType.APPLICATION_JSON)
-    public void updatePerson(@PathParam("billId") String Id,Billing bill ) {
+    public Response updatePerson(@PathParam("billId") String Id, Billing bill) {
         Billing existingBill = billDAO.getBillByInvoiceId(Id);
 
-        if (existingBill!= null) {
-             billDAO.updateBillingRecord(bill.getPayments(),bill.getCreditBalance(),Id);
+        if (existingBill != null) {
+            billDAO.updateBillingRecord(bill.getPayments(), bill.getCreditBalance(), Id);
+            return Response.ok().build();
+        } else {
+            throw new ResourceNotFoundException("No Bills found with ID:" + Id);
         }
     }
-    
+
     @DELETE
     @Path("/{billId}")
-    public void deletePerson(@PathParam("billId")String Id){
-        billDAO.removeBillingRecord(Id);
+    public Response deletePerson(@PathParam("billId") String Id) {
+        if (billDAO.getBillByInvoiceId(Id) != null) {
+            billDAO.removeBillingRecord(Id);
+            return Response.ok().build();
+        }
+        throw new ResourceNotFoundException("No Bills found with ID:" + Id);
     }
 }
