+++ src/main/java/com/mycompany/cw/resources/PrescriptionResource.java
@@ -11,11 +11,14 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import com.mycompany.cw.model.Prescription;
+import com.mycompany.exception.ResourceNotCreatedException;
+import com.mycompany.exception.ResourceNotFoundException;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.PathParam;
+import javax.ws.rs.core.Response;
 
 /**
  *
@@ -23,32 +26,50 @@ import javax.ws.rs.PathParam;
  */
 @Path("/Prescription")
 public class PrescriptionResource {
-    private PrescriptionDAO prescriptionDAO = new  PrescriptionDAO();
-    
+
+    private PrescriptionDAO prescriptionDAO = new PrescriptionDAO();
+
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public ArrayList<Prescription> getAllPrescriptions(){
-        return prescriptionDAO.getAllPrescriptions();
-    }
-   @POST
-   @Consumes(MediaType.APPLICATION_JSON)
-   public void addPrescription(Prescription prescription)
-   {
-       prescriptionDAO.addPrescription(prescription);
-   }
-   @PUT
-   @Path("/{prescriptionId}")
-   @Consumes(MediaType.APPLICATION_JSON)
-   public void updatePrescription(@PathParam("prescriptionId") String id,Prescription prescription){
-       Prescription existingPrescription = prescriptionDAO.getPrescriptionByID(id);
-
-        if (existingPrescription!= null) {
-           prescriptionDAO.updatePrescription(id,prescription.getDosage(),prescription.getPrescribedMedics(),prescription.getInstructions(), prescription.getDuration());
+    public ArrayList<Prescription> getAllPrescriptions() {
+        ArrayList<Prescription> prescriptions = prescriptionDAO.getAllPrescriptions();
+        if (prescriptions != null) {
+            return prescriptions;
         }
-   }
-   @DELETE
-   @Path("{prescriptionId}")
-   public void deletePrescription(@PathParam("prescriptionId")String id){
-       prescriptionDAO.removePrescription(id);
-   }
+        throw new ResourceNotFoundException("No Prescriptions Found");
+    }
+
+    @POST
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response addPrescription(Prescription prescription) {
+        if (prescription.getId() != null) {
+            prescriptionDAO.addPrescription(prescription);
+            return Response.ok().build();
+        }
+        throw new ResourceNotCreatedException("Cannot add Prescription");
+    }
+
+    @PUT
+    @Path("/{prescriptionId}")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response updatePrescription(@PathParam("prescriptionId") String id, Prescription prescription) {
+        Prescription existingPrescription = prescriptionDAO.getPrescriptionByID(id);
+
+        if (existingPrescription != null) {
+            prescriptionDAO.updatePrescription(id, prescription.getDosage(), prescription.getPrescribedMedics(), prescription.getInstructions(), prescription.getDuration());
+            return Response.ok().build();
+        }
+        throw new ResourceNotFoundException("No Prescription found with ID:" + id);
+
+    }
+
+    @DELETE
+    @Path("{prescriptionId}")
+    public Response deletePrescription(@PathParam("prescriptionId") String id) {
+        if (prescriptionDAO.getPrescriptionByID(id) != null) {
+            prescriptionDAO.removePrescription(id);
+            return Response.ok().build();
+        }
+        throw new ResourceNotFoundException("No appointments found with ID:" + id);
+    }
 }
